restructured modules and formatted
