+++ vicky/src/lib/database/entities/mod.rs
@@ -1,5 +1,9 @@
-mod task;
-mod lock;
+pub mod lock;
+pub mod task;
 
-pub use task::*;
-pub use lock::*;
+pub use lock::Lock;
+use rocket_sync_db_pools::database;
+pub use task::Task;
+
+#[database("postgres_db")]
+pub struct Database(diesel::PgConnection);
